Moving ConnectionSettings to core
