+++ gradle.properties
@@ -16,7 +16,7 @@ awsSdkVersion=2.17.19
 coroutinesVersion=1.3.9
 detektVersion=1.18.1
 jacksonVersion=2.11.1
-telemetryVersion=1.0.17
+telemetryVersion=1.0.20
 commonsMarkVersion=0.17.1
 jgitVersion=5.11.0.202103091610-r
