+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/cloudwatch/logs/insights/actions/QueryGroupAction.kt
@@ -5,7 +5,7 @@ package software.aws.toolkits.jetbrains.services.cloudwatch.logs.insights.action
 
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.project.DumbAware
-import software.aws.toolkits.jetbrains.core.credentials.ConnectionSettings
+import software.aws.toolkits.core.ConnectionSettings
 import software.aws.toolkits.jetbrains.core.credentials.activeCredentialProvider
 import software.aws.toolkits.jetbrains.core.credentials.activeRegion
 import software.aws.toolkits.jetbrains.core.explorer.actions.SingleResourceNodeAction
