+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/dynamodb/editor/DynamoDbVirtualFile.kt
@@ -5,8 +5,8 @@ package software.aws.toolkits.jetbrains.services.dynamodb.editor
 
 import com.intellij.testFramework.LightVirtualFile
 import software.amazon.awssdk.services.dynamodb.DynamoDbClient
+import software.aws.toolkits.core.ConnectionSettings
 import software.aws.toolkits.jetbrains.core.awsClient
-import software.aws.toolkits.jetbrains.core.credentials.ConnectionSettings
 
 /**
  * Light virtual file to represent a dynamo table, used to open the custom editor
