+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/schemas/ViewSchemaAction.kt
@@ -6,7 +6,7 @@ package software.aws.toolkits.jetbrains.services.schemas
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.project.DumbAware
 import icons.AwsIcons
-import software.aws.toolkits.jetbrains.core.credentials.ConnectionSettings
+import software.aws.toolkits.core.ConnectionSettings
 import software.aws.toolkits.jetbrains.core.credentials.activeCredentialProvider
 import software.aws.toolkits.jetbrains.core.credentials.activeRegion
 import software.aws.toolkits.jetbrains.core.explorer.actions.SingleResourceNodeAction
