+++ jetbrains-core/tst/software/aws/toolkits/jetbrains/core/credentials/AwsCredentialsExtensionsTest.kt
@@ -7,6 +7,8 @@ import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
 import software.amazon.awssdk.auth.credentials.AwsBasicCredentials
 import software.amazon.awssdk.auth.credentials.AwsSessionCredentials
+import software.aws.toolkits.core.credentials.mergeWithExistingEnvironmentVariables
+import software.aws.toolkits.core.credentials.toEnvironmentVariables
 import software.aws.toolkits.core.utils.test.aString
 
 class AwsCredentialsExtensionsTest {
